+++ schedulerx3-20240624/alibabacloud_schedulerx320240624/models/_update_executor_group_request.py
@@ -9,27 +9,35 @@ class UpdateExecutorGroupRequest(DaraModel):
         self,
         api_key: str = None,
         auth_type: str = None,
+        auto_scale: bool = None,
         cluster_id: str = None,
+        cms_workspace_id: str = None,
         description: str = None,
         id: str = None,
+        name: str = None,
         network: str = None,
         protocol: str = None,
         worker_type: str = None,
         workers: str = None,
+        xattrs: str = None,
     ):
         self.api_key = api_key
         self.auth_type = auth_type
+        self.auto_scale = auto_scale
         # This parameter is required.
         self.cluster_id = cluster_id
+        self.cms_workspace_id = cms_workspace_id
         self.description = description
         # ID。
         # 
         # This parameter is required.
         self.id = id
+        self.name = name
         self.network = network
         self.protocol = protocol
         self.worker_type = worker_type
         self.workers = workers
+        self.xattrs = xattrs
 
     def validate(self):
         pass
@@ -45,14 +53,23 @@ class UpdateExecutorGroupRequest(DaraModel):
         if self.auth_type is not None:
             result['AuthType'] = self.auth_type
 
+        if self.auto_scale is not None:
+            result['AutoScale'] = self.auto_scale
+
         if self.cluster_id is not None:
             result['ClusterId'] = self.cluster_id
+
+        if self.cms_workspace_id is not None:
+            result['CmsWorkspaceId'] = self.cms_workspace_id
 
         if self.description is not None:
             result['Description'] = self.description
 
         if self.id is not None:
             result['Id'] = self.id
+
+        if self.name is not None:
+            result['Name'] = self.name
 
         if self.network is not None:
             result['Network'] = self.network
@@ -66,6 +83,9 @@ class UpdateExecutorGroupRequest(DaraModel):
         if self.workers is not None:
             result['Workers'] = self.workers
 
+        if self.xattrs is not None:
+            result['XAttrs'] = self.xattrs
+
         return result
 
     def from_map(self, m: dict = None):
@@ -76,14 +96,23 @@ class UpdateExecutorGroupRequest(DaraModel):
         if m.get('AuthType') is not None:
             self.auth_type = m.get('AuthType')
 
+        if m.get('AutoScale') is not None:
+            self.auto_scale = m.get('AutoScale')
+
         if m.get('ClusterId') is not None:
             self.cluster_id = m.get('ClusterId')
+
+        if m.get('CmsWorkspaceId') is not None:
+            self.cms_workspace_id = m.get('CmsWorkspaceId')
 
         if m.get('Description') is not None:
             self.description = m.get('Description')
 
         if m.get('Id') is not None:
             self.id = m.get('Id')
+
+        if m.get('Name') is not None:
+            self.name = m.get('Name')
 
         if m.get('Network') is not None:
             self.network = m.get('Network')
@@ -96,6 +125,9 @@ class UpdateExecutorGroupRequest(DaraModel):
 
         if m.get('Workers') is not None:
             self.workers = m.get('Workers')
+
+        if m.get('XAttrs') is not None:
+            self.xattrs = m.get('XAttrs')
 
         return self