+++ schedulerx3-20240624/ChangeLog.md
@@ -1,3 +1,21 @@
+2026-05-28 Version: 1.7.1
+- Update API CreateExecutorGroup: add request parameters AutoScale.
+- Update API CreateExecutorGroup: add request parameters CmsWorkspaceId.
+- Update API CreateExecutorGroup: add request parameters IntegrationType.
+- Update API CreateExecutorGroup: add request parameters XAttrs.
+- Update API DeleteExecutorGroup: add request parameters Name.
+- Update API ListExecutorGroup: add response parameters Body.Data.Records.$.AutoScale.
+- Update API ListExecutorGroup: add response parameters Body.Data.Records.$.CmsWorkspaceId.
+- Update API ListExecutorGroup: add response parameters Body.Data.Records.$.IntegrationType.
+- Update API ListExecutorGroup: add response parameters Body.Data.Records.$.ResolvedWorkers.
+- Update API ListExecutorGroup: add response parameters Body.Data.Records.$.XAttrs.
+- Update API ListExecutors: add response parameters Body.Data.$.Status.
+- Update API UpdateExecutorGroup: add request parameters AutoScale.
+- Update API UpdateExecutorGroup: add request parameters CmsWorkspaceId.
+- Update API UpdateExecutorGroup: add request parameters Name.
+- Update API UpdateExecutorGroup: add request parameters XAttrs.
+
+
 2026-04-28 Version: 1.7.0
 - Support API ListExecutorGroup.